Separate notification handler from resourceForm.js
