+++ respa_admin/static_src/js/baseIndex.js
@@ -1,0 +1,7 @@
+import { initializeBaseEventHandlers }  from './base';
+
+function start() {
+  initializeBaseEventHandlers();
+}
+
+window.addEventListener('load', start, false);
