+++ respa_admin/static_src/js/resourceForm.js
@@ -25,7 +25,6 @@ let emptyDayItem = null;
 * Attach all the event handlers to their objects upon load.
 * */
 export function initializeEventHandlers() {
-  enableNotificationHandler();
   enablePeriodEventHandlers();
   enableAddNewPeriod();
   enableLanguageButtons();
@@ -50,16 +49,7 @@ export function getEmptyDayItem() {
   return emptyDayItem;
 }
 
-/*
-* Bind the event handler for closing notification to the elements (buttons)
-* since we cant do it directly in the html because of the scope
-* */
-function enableNotificationHandler() {
-  let notifications = document.getElementsByClassName('noti');
-  Array.prototype.forEach.call(notifications, (noti) => {
-    noti.getElementsByTagName('button')[0].addEventListener('click', () => noti.remove(), false);
-  });
-}
+
 
 /*
 * Set empty day and period variables.
