+++ respa_admin/templates/respa_admin/_base.html
@@ -59,7 +59,7 @@
 
     <!-- jQuery first, then Popper.js, then Bootstrap JS -->
     <script src="{% static 'respa_admin/styles-bundle.js' %}"></script>
-
+    <script type="text/javascript" src="{% static 'respa_admin/base-bundle.js' %}"></script>
     {% block js_extra %}
     {% endblock js_extra %}
 </body>
