+++ respa_admin/webpack.config.js
@@ -12,6 +12,7 @@ module.exports = {
     main: './static_src/styles/base.scss',
     styles: './static_src/js/styles.js',
     resourceForm: './static_src/js/resourceFormIndex.js',
+    base: './static_src/js/baseIndex.js',
   },
   output: {
     filename: '[name]-bundle.js',
